Install fetch-node and test leaderboard function
