+++ __tests__/ChaserShip.test.js
@@ -1,0 +1,9 @@
+import ChaserShip from '../src/entities/ChaserShip';
+
+test('It is a type of function', () => {
+  expect(typeof ChaserShip).toBe('function');
+});
+
+test('It is a type of function', () => {
+  expect(typeof ChaserShip).not.toBe('array');
+});
