+++ __tests__/SceneGameOver.test.js
@@ -1,0 +1,9 @@
+import SceneGameOver from '../src/scenes/SceneGameOver';
+
+test('should be type of function', () => {
+  expect(typeof SceneGameOver).toBe('function');
+});
+
+test('should be type of function', () => {
+  expect(typeof SceneGameOver).not.toBe('array');
+});
