+++ __tests__/SceneMain.test.js
@@ -1,0 +1,9 @@
+import SceneMain from '../src/scenes/SceneMain';
+
+test('should be type of a function', () => {
+  expect(typeof SceneMain).toBe('function');
+});
+
+test('should be type of a function', () => {
+  expect(typeof SceneMain).not.toBe('array');
+});
