+++ __tests__/ScrollingBackground.test.js
@@ -1,0 +1,9 @@
+import ScrollingBackground from '../src/background/ScrollingBackground';
+
+test('It is a type of function', () => {
+  expect(typeof ScrollingBackground).toBe('function');
+});
+
+test('It is a type of function', () => {
+  expect(typeof ScrollingBackground).not.toBe('array');
+});
