+++ __tests__/util.test.js
@@ -1,0 +1,7 @@
+import sortScore from '../src/util/sortScore';
+
+test('should return an array of objects', () => {
+  sortScore().then(data => {
+    expect(data).toBe('array');
+  });
+});
