+++ package.json
@@ -44,6 +44,7 @@
     "webpack-dev-server": "^3.11.2"
   },
   "dependencies": {
+    "node-fetch": "^2.6.1",
     "phaser": "^3.52.0",
     "regenerator-runtime": "^0.13.7"
   }
