--- tests/ChaserShip.test.js
@@ -1,9 +1,0 @@
-import ChaserShip from '../src/entities/ChaserShip';
-
-test('It is a type of function', () => {
-  expect(typeof ChaserShip).toBe('function');
-});
-
-test('It is a type of function', () => {
-  expect(typeof ChaserShip).not.toBe('array');
-});
